pyproject.toml
[tool.poetry]
name = "zunda-w"
version = "0.1.4"
description = "zunda-ASMR is speech to text to speech by voicevox(+α) and whisper ."
authors = ["OhkuboSGMS <ginger.os777146th@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "zunda_w" }]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9"
dataclasses-json = "^0.6.2"
py7zr = "^0.20.2"
pyunpack = "^0.3"
pydub = "^0.25.1"
requests = "^2.28.1"
srt = "^3.5.2"
classopt = "^0.1.9"
tqdm = "^4.64.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"
openai = "^0.27.2"
fire = "^0.5.0"
openai-whisper = {git = "https://github.com/openai/whisper.git", rev = "main"}
voicevox-engine-user-dict = { git = "https://github.com/OhkuboSGMS/voicevox_engine_user_dict.git" }
omegaconf = "^2.3.0"
ginza = "^5.1.2"
ja-ginza-electra = "^5.1.2"
torch = [
    { version = "2.0.0+cu118", platform = "win32", source = "torch-cuda" },
]
torchaudio = [
    { version = "2.0.0+cu118", platform = "win32", source = "torch-cuda" },
]
pathvalidate = "^3.1.0"
pandas = "^2.0.3"
langchain = "^0.0.300"
nltk = "^3.8.1"
pyannote-audio = "3.1.1"
flet = "0.19.0"
faster-whisper = {git = "https://github.com/SYSTRAN/faster-whisper.git",tag="v0.10.1"}
whisperx = {git = "https://github.com/OhkuboSGMS/whisperX.git",rev="94696797cdcf4b99da21f50b026ba68eff6300a8"}
#branch="feature/faster_whisper_0.10.0"}
transformers = "4.25.1"
sentencepiece = "^0.1.99"

[[tool.poetry.source]]
name = "torch-cuda"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.7.1"
black = "^23.7.0"
isort = "^5.12.0"
poethepoet = "^0.21.1"
pytest = "^7.4.3"


[tool.poetry.group.podcast.dependencies]
sfp-uploader = {git = "https://github.com/OhkuboSGMS/sfp_uploader.git"}
beautifulsoup4 = "^4.12.3"
atproto = "^0.0.40"
tweepy = "^4.14.0"
markdown = "^3.6"
mdx-linkify = "^2.1"
anthropic = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
zunda = "zunda_w.__main__:_main"
zunda_gui ="zunda_w.gui.main:__main__"
[tool.poe.tasks]
_black = "black ."
_isort = "isort ."
format = ["_isort", "_black"]